add model interface, services
